Minor update to current class parser
I noticed that there was some buffering issue with files not closing
well in python and then being reused within same script. So, I just cut
out the writing to file piece and keeping all in RAM now - these are not
big files, shouldn't be any problem. This may also be part of the reason
for missing classes/broken foreign keys (beyond 1/1).