Claims (re)parser for 2005+
Since we already tested the claims parser from all ends and there seems
to be truncated text in current version from PatentsProcessor, let's use
this code to re-upload. The claims_parser.py is the wrapper to run
through command line. Do this as the very last step in creating testDB.